Preliminary testing and CI infrastructure for the Github repository
This required several changes accross the various packages, since the assumption
that the projects are built and tested in a different type of environment was
pretty ubiquitous.

The given `Dockerfile` allows one to build the projects and run the tests.
It pulls in various dependencies, from apt as well as github repos, some are
C++ libraries, others are haskell tools.

This image is then used by `.github/workflows/ci.yml` to build the packages in
this repository as well as test them, making sure that we generate some files
at the right time along the way.

This isn't quite enough to get all tests to pass. More on that in the
pull request.