Marshall IOBuf to Lazy ByteString
Summary: Marshall a `folly::IOBuf `to Haskell lazy `ByteString`. This is very much a draft/experimental change which I thought might be useful (I realized I would like it for my project). Very open to optimizations and suggestions especially given unfamiliarity with the nuts and bolts of GHC and the FFI, and relative costs of calling different functions in Haskell/C. Unlikely to be useful for small `IOBuf`s, but for chains of very large buffers this might be good to have. Not particularly attached to this implementation and willing to abandon if impractical.

Reviewed By: watashi

Differential Revision: D24512663

fbshipit-source-id: a82314093e81359633a7d54b4aebb0da55e2419b