Avoid MonadFail in generated code
Summary:
Just trying to avoid situations in which the target monad does not have a `MonadFail` instance.
I can see an argument for sticking to `fail` in certain cases, but custom implementations of `MonadFail` are rare - when/if the requirement to use `fail` arises we can revisit this code generator.

zilberstein simonmar please confirm that we don't really need `fail` here.

Example compile error with GHC 8.8:
```
/data/sandcastle/boxes/eden-trunk-hg-fbcode-fbsource/fbcode/buck-out/dev/gen/titan/iris/extif/IrisCommon-hs2-IrisCommon.thrift/gen-hs2/IrisCommon/IrisThriftService/Service.hs:63:17: error:
    • Could not deduce (Prelude.MonadFail ((->) Int.Int32))
        arising from a use of ‘Prelude.fail’
      from the context: Thrift.Protocol p
        bound by the type signature for:
                   respWriter' :: forall p a.
                                  Thrift.Protocol p =>
                                  Proxy.Proxy p
```

Reviewed By: josefs

Differential Revision: D25271099

fbshipit-source-id: 62e8fe98571029e752b983edb7f29050fcbf7edc