Support producing proper source distributions
Where each package can be tested on its own (and therefore embeds all the
appropriate files), via a source distribution. The Makefile changes
introduce rules for copying all the right files in the right places.

The compiler's test infrastructure is appropriately updated to accomodate for
the new scenario, where we can't just hardcode the repository's structure and
need to be both a bit more careful in how we set things up for the tests, and
more lenient when it comes to comparing the expected output and what the test
actually produced, occasionally stripping some directory name from paths.