thrift-lib: introduce socket channel and accompanying tests (#11)
Summary:
This is a preliminary implementation of socket-based channels in `thrift-lib`, which comes with similar tests as other channels.

simonmar I'm sure there will be things to tweak, but with this branch as it is now, I can reliably and repeatedly get the tests to pass in CI _and_ on my machine.

Pull Request resolved: https://github.com/facebookincubator/hsthrift/pull/11

Reviewed By: zilberstein

Differential Revision: D26072950

Pulled By: simonmar

fbshipit-source-id: 94f9cabb7106a7594300a25f88169514810df351