Cleanly separate the generic layer from CPP/fbthrift
Summary:
We were pulling in the priority constants from fbthrift, so now I'm
defining those in RpcOptions.thrift. They don't necessarily have to be
in sync with the fbthrift definitions, but I've left it that way for
now, using static_assert() in the C++ code to catch divergence (which
I guess is unlikely to happen, but it doesn't hurt to be paranoid).

Differential Revision: D26275610

fbshipit-source-id: 3e9af7196508d5f458b12793d338e79a090480f8