Cleanly separate the generic layer from CPP/fbthrift (#21)
Summary:
Pull Request resolved: https://github.com/facebookincubator/hsthrift/pull/21

We were pulling in the priority constants from fbthrift, so now I'm
defining those in RpcOptions.thrift. They don't necessarily have to be
in sync with the fbthrift definitions, but I've left it that way for
now, using static_assert() in the C++ code to catch divergence (which
I guess is unlikely to happen, but it doesn't hurt to be paranoid).

Reviewed By: zilberstein

Differential Revision: D26275610

fbshipit-source-id: 7ee9d6b8ecccc4db855c3e17fd71eb9a072f05fa