Enable runtime bespokes for Tao deserialization
Summary:
Integrate StructDict into one chunk of Tao (deserialization). The VariantControllers are updated to allow for creation of a map from a RuntimeStruct definition, and the setters now accept indices. Includes some minor refactoring of IntishCast.

Further Tao integration is coming, but this was the most sensitive set of changes, so up as a separate diff.

Differential Revision: D27628853

fbshipit-source-id: 64ba18c7c2ca102e39ebabab85bb889c1c4f3434