Remove qualified access to own module name before typechecking
Summary:
We want to cope with foo.thrift that defines a named type 'bar' and refers to it as 'foo.bar' (instead of plain 'bar').

In lenient mode detect these cases and strip the leading 'foo.' before typechecking.

( Attempts at a more subtle approach did not work, so this diff uses a simple rewrite. )

Reviewed By: nhawkes

Differential Revision: D27499215

fbshipit-source-id: ea55536d55d43d49594838ab1db280d39883a3d8