Strip / avoid prepending envelopes on response path in clients (SingleRequestSingleResponse only)
Summary:
Currently we prepend the legacy envelope in Rocket layer (which is our default transport at this point), which is quite expensive.

This diff changes `ClientRecieveState` to expose only the body part that doesn't have an envelope (i.e. `SerializedResponse), which allows us to stop needing to prepend the envelope.

This diff changes only the SingleRequestSingleResponse path, we still do prepend it in the Stream/Sink cases and at the point of integration with languages.

We need to later clean this up even more to stop using legacy envelope in the rest of cases.

Reviewed By: iahs

Differential Revision: D28297806

fbshipit-source-id: 0e61f6a3018637802c9c9dddd6e4cf12d4d80ed7