Add sendException to ResponseChannelRequest in haskell
Summary:
Add a sendException path to ResponseChannelRequest so that implementations (rocket) can infer T_REPLY or T_EXCEPTION when receiving messages which don't contain the prologue that includes the type.
Default implementation of sendException forwards to sendReply which will work until Rocket requests responses without the envelope which will happen at the end of this stack of diffs.

Reviewed By: yfeldblum

Differential Revision: D28815049

fbshipit-source-id: e11cdd25e288613d4a971d7d6a7fc24773794b3a