HsOption + std::optional
Summary:
I have a need for an HsStruct that backs `std::optional`. I want to be able to use `withCxxObject` and `withDefaultCxxObject` to take advantage of avoiding hitting the c++ heap when possible, so `HsMaybe` doesn't cut it.

`std::optional` is tricky because the size and alignment of the object is dependent on the underlying type `T` being held within the `optional`. The solution presented here is to use `TH` and some macros to generate the desired instances.

There's one gross piece of coupling between the generated Dummy cpp object and needing to know the exact name in the `hsc` `#{size ...}` call. Since `hsc` runs before `TH` I didn't know of a way to generate the cpp object name.

Reviewed By: watashi

Differential Revision: D28877285

fbshipit-source-id: dd94061bda1acfb059a7cba6b6238f486f026c77