HsArray Constructible instances
Summary:
In order to use `HsArray` and `HsList` under `withCxxObject` and `withDefaultCxxObject` we need Constructible instances. We unfortunately can't implement generic FFI calls that would work for all underlying types, so instead use some more TH and macro magic to generate the desired instances for us.

This opens the question of how to send a list of objects down to cpp to be constructed into a `std::vector`. I see 2 options:
- Across all elements, use `with` or `withCxxObject` on it and call an `emplace` FFI to move the data into the vector.
- Construct a void* of the right size, filling in each value with `constructValue` in a bracketed fashion, then make one FFI call construct the std::vector from that array.

This sounds like a classic Ninja coding question trade-off as the second one requires a stack as deep as the number of elements to pass along. I've implemented the first option here since it was easier to write. If we find that this is a bottleneck we can optimize conditionally on the size of the vector later on.

Reviewed By: simonmar

Differential Revision: D29204072

fbshipit-source-id: b04c9209e7174894714c1caeced35caedc864852