Fix compiler bug for empty services
Summary:
The generated Haskell code for empty services would not compile due to trying to export the constructors of empty GADTs.

Fix this by not exporting the constructors for empty services — also remove the "where" from the data declarations for tidiness.

Also add a build target to test that these Thrift files compile properly now.

Reviewed By: watashi

Differential Revision: D29393825

fbshipit-source-id: e290d13fe282c2cb3d7b521c6fb286b35ab49c9a