Fix compiler bug for empty services
Summary:
The generated Haskell code for empty services would not compile due to trying to export the constructors of empty GADTs.

Fix this by not exporting the constructors for empty services — also remove the "where" from the data declarations for tidiness.

Also add a build target to test that these Thrift files compile properly now.

Differential Revision: D29393825

fbshipit-source-id: 10e10c7ca767900c69e5f0a29dc3a35ef55b952f