Define and expose peekHsEitherWith and peekHsPairWith
Summary:
Define and expose `peekHsEitherWith` and `peekHsPairWith`.
`peekWith` only allow us to pass a function to selectively peek the
right/second case. `peekHsEitherWith`/`peekHsPairWith` will allow us to
customize how to peek both the left/first and right/second case.

This is useful if we are only interested in a subset fields of left/first
or left/first is not an instance of Storable.

Reviewed By: ahaym

Differential Revision: D29851653

fbshipit-source-id: 1d8dcc1cdd8ce2dc892b8ff1c535135ee8178405