Refactor the sendRequest methods into one
Summary:
This refactors the common parts of `sendOnewayRequest` & `sendRequest`
into one, by leveraging an enum in choosing the request-direction. This allows
us to iterate on how we compose the actual request in one place rather than in
two (where we'd have a higher chance of divergence). It also aims to make it a
bit easier to follow the logic.

Reviewed By: simonmar

Differential Revision: D31790274

fbshipit-source-id: bb4b36fe8170e13287b1e09da4bb47714150f536