Add support for thrift middleware
Summary:
Thrift has a concept of middleware (processor-handlers), which are installed and called upon every thrift request. Installing a handler allows you to hook into thrift to for example setup context propagation. It primarily allows you to modify the headers or anything in the request in itself. There's a bunch of different context we propagate around for e.g. tracing purposes and enforcement purposes internally (such as for privacy).

This diff
1. Modifies the sendRequest methods to create a context-stack and apply all middleware we have installed
2. Modifies the `HsCallback` to hold a unique-ptr reference to the context-stack in order to ensure that its lifetime is larger than the request, as part of the context is now embedded / installed in the request (I'm guessing for perf reasons to avoid unnecessary copies)

Reviewed By: mshneer

Differential Revision: D31685183

fbshipit-source-id: 71204c4db4bd47f0907c8041bb16eca24300550a