Better support for building hsthrift deps with getdeps.py
getdeps tends to over-build dependencies for hsthrift, not realizing
they are resolved by system packages. This leads to longer build times
and a more complicated linker environment.

In projects like folly, this is solved by deleting e.g. openssl from the
manifest, to make it an implicit system dep. We can't do that, but we
can just use --no-deps pointwise build of the dependencies.

Benefits:
- we get to use getdeps.py, which simplifies how to build
- and doesn't need sudo

We still have to teach cabal about LD_LIBRARY_PATHS and PKG_CONFIG_PATH
as in the old style, but we avoid requiring root privs to install now.