Better support for building hsthrift deps with getdeps.py (#60)
Summary:
getdeps tends to over-build dependencies for hsthrift, not realizing they are resolved by system packages. This leads to longer build times and a more complicated linker environment.

In projects like folly, this is solved by deleting e.g. openssl from the manifest, to make it an implicit system dep. We can't do that, but we can just use --no-deps pointwise build of the dependencies.

Benefits:
- we get to use getdeps.py, which simplifies how to build
- and doesn't need sudo

We still have to teach cabal about LD_LIBRARY_PATHS and PKG_CONFIG_PATH as in the old style, but we avoid requiring root privs to install now.

To use:
./new_install_deps.sh
make as before

This PR updates the ci.yml to be green, and uses -j2 to build folly , bringing build times back in line with old ./install_deps.sh

Pull Request resolved: https://github.com/facebookincubator/hsthrift/pull/60

Reviewed By: nhawkes

Differential Revision: D33843787

Pulled By: simonmar

fbshipit-source-id: c3a0018a2608532da1aff5255e0aa713d5240023