Add support for installing additional packages from cmd line args (#61)
Summary:
This extends the deps installer with enough smarts to install additional packages on the command line.
We do this so that Glean can ask the hsthrift build_deps.py to also solve the rocksdb source dependency, which is now required.

No changes to hsthrift itself, but this will unblock glean. Updates the rocksdb manifest we use to point at git.

Pull Request resolved: https://github.com/facebookincubator/hsthrift/pull/61

Reviewed By: nhawkes

Differential Revision: D33889630

Pulled By: simonmar

fbshipit-source-id: 3050f2c205b454168970b08101314c938f5d995f