ignore structural annotation for service functions
Summary:
## Context

In si_sigma, we don't generate all thrift symbols because there are too many unused. Instead, we run a analysis to understand what are thrift symbols actually used, and
when hsthrift compiles, it first performs "prune".

## The problem

During pruning process, all structural annotations should be ignored, but the one for services functions are not. So when performing typecheck, it will emit decl not found problem and failed the compilation

## The impact

We were able to mitigate this by have a dummy reference (https://fburl.com/code/y6s2wa67), which makes hsthrift keep the annotation decls. However, we have a codemod which clean unused imports, which will see dummy references as "not used", meaning, this codemod is dead effectively.

ref post: https://fb.workplace.com/groups/333784157210625/posts/982888848966816

Reviewed By: nhawkes

Differential Revision: D34029528

fbshipit-source-id: 084b148ed79ac1a72452eb85e2b626ba583de7b2