Make thrift-cache rule more robust (#72)
Summary:
The output of list-bin is sensitive to -v flags. If we use -vnormal or
above we might get spurious "Up to date" output on the cached key, which breaks the build.
Use -v0 on the cache rule to force cabal to just give us the path. This should be more resiliant if users set other flags.

Pull Request resolved: https://github.com/facebookincubator/hsthrift/pull/72

Reviewed By: pepeiborra

Differential Revision: D34712711

Pulled By: simonmar

fbshipit-source-id: 42efbd272a3bd3ca1a6cfc445696af1064c3b90a