Suppress -Wterminate for createDynamic is marked noexcept but throws (intentionally)
This is a slightly worrying looking warning, but we are intentionally
throwing (and std:terminating) in this noexecpt method. So mark it with
the pragma and make the warning go away

Without this you get a warning along the lines of "this will always
terminate and throws!" -- which we know.