Suppress -Wterminate for createDynamic is marked noexcept but throws … (#75)
Summary:
…(intentionally)

This is a slightly worrying looking warning, but we are intentionally
throwing (and std:terminating) in this noexecpt method. So mark it with
the pragma and make the warning go away

Without this you get a warning along the lines of "this will always
terminate and throws!" -- which we know.

Pull Request resolved: https://github.com/facebookincubator/hsthrift/pull/75

Reviewed By: DylanZA

Differential Revision: D34752155

Pulled By: simonmar

fbshipit-source-id: f5dfcb8906fbb93c9540b6cd25e6f6cbdc6f7fdf