Make the content of RequestContextPtr always a nonnull shared_ptr
Summary:
Make the content of RequestContextPtr always a nonnull shared_ptr, this
simplifies a lot of edge cases handling. A null and empty ReqeustContext
had little differences in most use cases. We only had null in tests
and some tools anyway, it should always be nonnull when `saveContext` is called
in prod services.

This also ensured that a Haskell `ReqeustContext` has a constant C++ `uintptr_t`
to identify its uniqueness and equality. We may use it to implement `Hashable`
and `Eq` instance so it can be used a hashtable key at runtime.

Reviewed By: josefs

Differential Revision: D35760371

fbshipit-source-id: 1fbf58b31f58527052475f2de876017262d006b3