fb-util: Include Executor.cpp
On systems with LD_BIND_NOW set or ghc with RTLD_DEFAULT linking enable,
compilaton of fb-util fails due to missing symbol
`common_hs_releaseGlobalCPUExecutor`. This is defined in Executor.cpp,
so we need to include it into the compilation.

Now I tried to add Utils/Executor.cpp and still got the errors. When
I moved the file to cpp/Executor.cpp and included it, it worked. Now I
have no clue why that is, and i tried to clean build artifacts but still
had issues when Executor.cpp is not in cpp/. So I left it, without
actually knowing _why_ this is.