fb-util: Remove unused Util.Dll
Util.Dll is exports functions that don't exist in RTS of GHC 8, but
GHC 8 is required for compilation, hence on systems with a RTLD_DEFAULT
enabled ghc or LINK_BIND_NOW=1 set, we get linker errors. Luckily
our main downstream Glean does not use Util.Dll, so we can just remove
this.

Alternatively we can either require GHC 9 (and upgrade all our
dependencies) or remove setHighMemDynamic and other functions that aren't
available in GHC 8 (such as loadNativeObj and unloadNativeObj)