fb-util: Remove unused Util.GFlags
c9fb3a4 (Initial Commit) already removes the compilation of
Utils/GFlags.cpp, but we still include Utils/GFlags.hs which depends
on GFlags.cpp exported functions. This will fail at runtime, and on
systems with LD_BIND_NOW=1 or an RTLD_DEFAULT enabled ghc this leads
to link errors on build time. So remove it.