Add a callstack to UnknownDatabase exceptions
Summary:
malanka noticed that we are getting `UnknownDatabase` errors in Production that arise after deleting a DB. It's almost certain that they are being raised by `Catalog.getEntry`, but there are >50 transitive call sites which makes guessing a bit hard.

This diff adds a very simple rendered call stack to the `UnknownDatabase` exception, which will be included in the server exception logs.

Reviewed By: simonmar

Differential Revision: D37754521

fbshipit-source-id: acd613f5b9a92082048237f8082910b4f1d6b6c9