Make cpp/wrap.h header only
This is used by some C++-only libraries and binaries (e.g., the C++ indexer) which we don't really want to depend on Haskell libs. We should probably move the C++ FFI bits into a separate lib but making it header-only for now is easy and might be the right thing to do anyway (the address comparisons we used to do might actually be dodgy with dynamic linking).