Reexport Control.Concurrent.STM
Summary:
This means that `Util.STM` is a complete replacement for `Control.Concurrent.STM`, and `common/hs/util:stm` can often be used in place of the `stm` package.

The motivation is to make it easy to replace `atomically` everywhere by simply:

1. Replacing the `stm` package with `common/hs/util:stm`
2. Substituting the import

Doing it this way ensures that no `Control.Concurrent.atomically` is accidentally left behind.

Reviewed By: bochko, malanka

Differential Revision: D42139846

fbshipit-source-id: cae8119fdbf6e19f731870c9793bb597be3fec69