Fix performance bug with text slicing
Summary:
`slice` is not a library function but it is trivially implementable as a function composition `take . drop`. Unfortunately, due to a perf bug in the `text` library, this implementation has terrible performance. More information here:

https://github.com/channable/haskell-string-slicing-benchmarks

Here we extract a `slice` library function with a custom implementation for the affected versions of the `text` library, and update `glean/tools/derive` to use it, resulting in a massive speed-up.

I also added a linter rule to catch the troubled code and suggest using `slice` instead. The linter found one ocurrence in `Glean.Glass.Util` so I took the chance to fix it too

Reviewed By: phlalx

Differential Revision: D43446269

fbshipit-source-id: 1e18debad70f598ee11ab7f65225189385f517e8