write: halve Max RSS and improve perf
Summary:
We are deserializing the JSON facts only to immediately serialize them again in the leaves; this is needed to put the batch in the encoding expected by the Write server, and it uses up a lot of memory for large files since the Aeson `Value` representation is not memory efficient.

A cheap improvement is to do the JSON manipulation in Folly, which is hopefully more memory efficient and faster than Aeson.

Reviewed By: simonmar

Differential Revision: D44415801

fbshipit-source-id: a972e49ea0f9b0f29569a09acc80fefb7149525d