Fix a memory bug
Summary:
EliasFano needs to be able to read 8 bytes past the end of the data,
so we have to copy the bytes we get from RocksDB before we can read
the contents of the set.

I have actually seen this crash, and it definitely gets flagged by
ASAN. Seems unlikely that this could be responsible for bad data, but
it could cause server crashes.

Reviewed By: pepeiborra

Differential Revision: D46729613

fbshipit-source-id: b74a1bc09a6a01a6b53e978e110ce1462a9485d5