Use HsLenientText for any text from storage in ligen
Summary:
Use HsLenientText for any text from storage in ligen.
It's always that the data from storage may be corrupted and not valid utf-8
encoded. It's safer to use lenient decode in this case.
We may want to use `TextAndByteString` if we care about the original corrupted
data, but that will make the api less user friendly and we may simply mark
the type as `binary` instead.

Reviewed By: babyLiyuan

Differential Revision: D49473049

fbshipit-source-id: df034e1b261b2d1ddbcaec36208a20954d7563f8