Modernize the typechecker monad and make it abstract
Summary:
This is a no-op diff that just reorganizes the code a bit.

The typechecker monad is made abstract. This allows for removinging the specialized `collect` function in favour an `Applicative` instance. Similarly, `orT` and `emptyT` are replaced by an `Alternative` instance.

I've also removed `mapT` and `mapE` as they were just `traverse`.

Reviewed By: simonmar

Differential Revision: D53135425

fbshipit-source-id: a050c73729ef51ad56ea85126e8dbf0cc15a9583