Get Haskell Thrift compiler to recognize typedef'd enum constants
Summary:
There's been a long-standing bug in this compiler where it won't recognize enum constants correctly if the constant is referred to by a typedef. See D58894832 for an example test case.

I think this diff fixes the bug, but it needs a pretty tight review, since I haven't written Haskell in a very long time. Don't hesitate to call out anything wrong or non-idiomatic or etc.

Also, I found that there's some enum constants that are accepted by the compiler that are not described in [the spec](https://www.internalfb.com/intern/staticdocs/thrift/docs/idl/#constants). Specifically, referring to just the enum value, without the enum name (and with or without a qualifying module). I'm not sure how common these are in practice, or whether they're accepted by the standard Thrift compiler; does anyone happen to know anything more about those formats? Or maybe that link is wrong?

Reviewed By: simonmar

Differential Revision: D58906253

fbshipit-source-id: c712113b1e993ec7b94e4b047e380ba9fcda7ae2