Build fast_float from source (#145)
Summary:
Folly requires a more up-to-date fast_float than the one in Ubuntu 24,
so CI is currently using Ubuntu 25. But a better solution is to build
fast_float from source, and then we can revert the CI to use Ubuntu 24,
so that we're back on an LTS release.

Pull Request resolved: https://github.com/facebookincubator/hsthrift/pull/145

Reviewed By: malanka

Differential Revision: D71552122

Pulled By: iamirzhan

fbshipit-source-id: 962b1208289681cbcd658080fd28856244a932a4