Import Data.Default for constants
Summary:
When working on D73144231, I ran into a bug in the Haskell Thrift
codegen. It occurs when you have a thrift file that defines a constant struct,
but doesn't define any new structs itself. (Likely also happens with unions;
didn't try it though.)

This diff contains a minimal repro of the issue, in `no_structs.thrift`, and
also a fix.

## Problem

In the generated Haskell for `no_structs.thrift`, (P1798036002) the definition
of the constant starts with the default instance of `Foo`:
```
fooConst = (Default.def :: Foo.Foo){Foo.foo_foo2 = Prelude.True}
```

But this does not compile (P1798054485):
```
buck-out/v2/gen/fbcode/27a1851cc8046c56/common/hs/thrift/tests/if/__no_structs-hs2-no_structs.thrift__/out/gen-hs2/NoStructs/Types.hs:21:13: error:
    Not in scope: ‘Default.def’
    No module named ‘Default’ is imported.
   |
21 | fooConst = (Default.def :: Foo.Foo){Foo.foo_foo2 = Prelude.True}
```

Comparing it to `foo.thrift`, which does compile, we see that it's missing the
following line:
```
import qualified Data.Default as Default
```

## Fix

I want to preface this with the fact that my Haskell is not very good and
probably unidiomatic, so feel free to nitpick to your heart's content.

Anyways. That import seems to be generated by `GenStruct.hs`, `GenUnion.hs`,
etc. There's a corresponding `GenConst.hs`, and there's a `genConstImports`
function that looks really promising, so I just tacked on a `baseImports` set,
copying the existing code.

Reviewed By: pepeiborra

Differential Revision: D73794875

fbshipit-source-id: 314aa4b297cb062445b3aab13fc50347f7de4bef