Bundle folly sources & build without getdeps
The goal is to enable `cabal install glean`, i.e. build Glean from
Hackage entirely under the control of Cabal, no external build systems.

On a recent enough Linux distro, the only dependency that still requires
getdeps is folly. In this PR I'm adding support for bundling the sources
of folly into a Cabal package so that it can be built as a regular
package.

There's no literal copy of folly in the repo, instead to build the Cabal
package we do

```
make setup-folly
```

which clones folly under `folly/`, updates to the correct revision from
`build/deps`, and runs `cmake` to generate `folly-config.h`. We also
download `fast-float` and unpack it under `folly/`.

Then `cabal sdist folly` will pack up the bundled sources as usual.

There's a Cabal flag `bundled-folly` which enables use of the bundled
sources, otherwise we just depend on folly from `pkg-config` which works
when we build from `getdeps`, so the existing `getdeps` build will
continue to work provided we disable `bundled-folly`. The provided
`cabal.project` files do this automatically, so CI continues to work.