'support' parsing bidi sink
Summary:
Unbreak the LLM Predictor build to allow adding `bidi` support.

This diff "fixes" it by making bidi stream parse in the Haskell compiler. To make this work, the following two cases need to be added to grammar:
- `sink<SinkT throws?>, stream<StreamT throws?>`
- `FirstResponseT, sink<SinkT throws?>, stream<StreamT throws?>`

Both of these are left unimplemented. They just parse the `stream` info and throw out the `sink`.

This should be ok because it's unlikely that Hs compiler will ever receive new features.

Reviewed By: CatherineGasnier

Differential Revision: D88114148

fbshipit-source-id: 554433e570d5e5aad1c268d4a0c47a807b5c52e9